+++ src/index.ts
@@ -1,2 +1,3 @@
 import './out-port-picker';
 import './in-port-picker';
+import './scale-picker';