+++ CHANGELOG.md
@@ -9,3 +9,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added out-port-picker & in-port-picker
+
+- Added scale-picker